third_party/blink/web_tests/platform/generic/html/tabular_data/tr_cells-expected.txt

Test behavior of the HTMLTableRowElement cells attribute in cases where there is unusual nesting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkCellNesting("col") is 0
PASS checkCellNesting("colgroup") is 0
PASS checkCellNesting("div") is 0
PASS checkCellNesting("form") is 0
PASS checkCellNesting("script") is 0
PASS checkCellNesting("table") is 0
PASS checkCellNesting("tbody") is 0
PASS checkCellNesting("tfoot") is 0
PASS checkCellNesting("thead") is 0
PASS checkCellNesting("tr") is 0


PASS checkCellNesting("td") is 1
PASS checkCellNesting("th") is 1


PASS checkHeaderCellNesting("col") is 0
PASS checkHeaderCellNesting("colgroup") is 0
PASS checkHeaderCellNesting("div") is 0
PASS checkHeaderCellNesting("form") is 0
PASS checkHeaderCellNesting("script") is 0
PASS checkHeaderCellNesting("table") is 0
PASS checkHeaderCellNesting("tbody") is 0
PASS checkHeaderCellNesting("tfoot") is 0
PASS checkHeaderCellNesting("thead") is 0
PASS checkHeaderCellNesting("tr") is 0


PASS checkHeaderCellNesting("td") is 1
PASS checkHeaderCellNesting("th") is 1


PASS successfullyParsed is true

TEST COMPLETE
